+++ components-compose/src/main/java/uk/gov/hmrc/components/compose/ui/theme/Color.kt
@@ -32,6 +32,7 @@ val HmrcGrey3 = Color(0xFFF3F2F1)
 val HmrcPink = Color(0xFFCA2B75)
 val HmrcYellow = Color(0xFFFFBF47)
 val HmrcAlwaysBlack = Color(0xFF0B0C0C)
+val HmrcDonutBlue = Color(0xFF003078)
 
 // Dark theme colors
 val HmrcBlackDark = Color(0xFFFFFFFF)